Make toggling the HOLE flag reversible
Closes-bug: lp-699483

When toggling the HOLE flag, the Thickness is no longer changed.
Instead, use PIN_SIZE() to select from Thickness or DrillingHole
accordingly, and use that throughout.

Mask size is handled thusly: For tented vias, the mask is opened up to
just over the hole size.  For untented vias, the mask gap is held
constant.

Also, various minor changes to ensure that untented holes are drawn
correctly.