hid/gtk: Allow zooming out past the board size (up to 1/10 viewport size)
The 1/10 viewport size should be the same as the Lesstif HID allows. The
purpose of the limit is to avoid zooming in so small the PCB is lost as a
tiny dot, and also to avoid coodinate overflows when representing screen
coordinats in PCB Coords. (Since we switched to nanometers, this could
potentially raise its head more readily until we move to 64bit integers).

The explicit pan fixup added to ghid_view_zoom_fit() is required as
pan_common() no longer clamps the view back to the origin.