Introduce GtkPcbLayerSelector widget -- not used yet
This widget will replace the layer-selection buttons in the
Gtk GUI, as well as the layer selection and visibility-toggling
parts of the menu.

This is to make layer manipulation more consistent and
more keyboard-accessible.

In future, it would be good to update this widget with context
menus with things like "hide all but this layer".