gtk: remove call to ghid_layer_buttons_color_update
The call to ghid_layer_buttons_color_update() in config_read()
is now unnecessary; the Gtk layer selection widget does not
exist, nor do the PCB struct's colors need to be synced with
the ones in the Settings struct.

So this call makes no sense. Oh, and it causes a segfault.

Also: remove color-changing debug code from gui-top-window.c