hid/ps: Fix inaccurate comment and twiddle some whitespace
The whitespace changes in the if statement were from an earlier revision
which actually had functional changes - but I prefer this way, so I'm
committing it anyway.

Also adds some != 0 to the end of strcmp tests, as I believe this aids
clarity by reminding readers strcmp returns 0 for a match.