Support adding/deletion of layers in GtkPcbLayerSelector
If you add a layer to a GtkPcbLayerSelector, and its ID is already
in the selector, it will update the layer instead of adding a new
one. This way, we can update the layer selector by:

  1. Deleting all recently-deleted and non-copper layers.
  2. Re-adding all layers (including new ones)
  3. Re-adding all non-copper layers (so they go to the end)

The old layer selector maintained MAX_LAYER + n layers (where n was
the number of non-copper layers), and showed/hid the gui widget to
managed deletion and adding of layers. The new one has no notion of
MAX_LAYER, nor does it care whether a layer is copper or not. :)