Better handling of unplated vias
The attached patch adds checks in the geometry intersection functions to
no longer treat unplated vias (mounting holes) as conducting. The
implementation in the patch will act as if the unplated via doesn't
intersect with metal objects at all, which is possibly not what is
desired. Should the conductivity check happen on a higher level from the
geometry intersection routines? A case that would require this is not
obvious to me.

Closes-bug: lp-699499