Replace gtk UI manager with GHidMainMenu widget
This commit replaces the old UI manager with a customize gtk
widget that builds menus directly from a resource tree. This
eliminates the translate-to-XML step, the ugly hacks used to
access individual actions, and all the associated manual
memory management.

This will also give us the ability to have more dynamic menus,
in particular layer lists without maximum capacities.

Layers and route styles are still not hooked into the menu.
This means that those accelerators DO NOT WORK. (This will
be fixed in a later commit.) Checkboxes have been replaced
with radio buttons where appropriate. There are now tearoffs
on the context-menu's submenus.

Other than that, there should be no user-visible changes. ;)