DrawLayer(): Pass true to clear_pins of DrawLayerCommon
This should not affect any rendering. If the check planes flag is set,
we should only draw polygons, nothing more. Only a handful of cases draw
layers via this code-path, and in all but the exporter cases, they are
not called if CHECKPLANES is set.

The exporters which call DrawLayer explicitly clear the CHECKPLANES flag
(since commit 3c01bd38cb59922692408b71cd5d77892dbe6ade), so nothing
should be affected by this argument change.