Block signals in ghid_main_menu_update_toggle_state()
The function ghid_main_menu_update_toggle_state() is supposed to
sync the menu checkboxes to the actual state of pcb, by checking
the flags given in the resource file.

If this function is emitting signals and changing pcb's state,
this is a bug. Fortunately, this appears not to be the case, so
the effect of this commit is to eliminate one more use of
ghidgui->toggle_holdoff.