hid/gtk: Don't pack the drawing area in a GtkViewport widget
(For both the main window and the pinout preview). The GtkViewport
widget is designed to add scroll bars to a widget which doesn't have
native scroll capabilities. We are handling out own scroll-bars, so
the only gain we had from the GtkViewport we added was a shadow
around the widget.

If we decide we want the shadow back, a more appropriate widget to
pack the drawing area with might be a GtkFrame.