Add configure selection of 32/64 Coord type
  --enable-coord64        Force 64-bit coordinate types
  --enable-coord32        Force 32-bit coordinate types

Defaults to "long" as before, but if you give one of the above,
the type changes to a suitable 32-bit or 64-bit type.  Note
that this is only guaranteed to be the size you choose if you
have <stdint.h> which most OSs provide, else the "int" and "long long"
types are used instead.