Fix text scaling exceeding Coord max.
The old way of scaling text was to multiply by scale/100 but this could
easily overflow on reasonably sized boards with a 32-bit Coord type.
The new code scales by (double)scale/100.0 instead.  Since we don't
store scaled values in the PCB file, a loss of precision won't matter,
but a double has 53 bits of precision - in nanometers, that's a board
about the size of North America.

Closes-bug: lp-832451