hid/gtk: Avoid double-destroying the file-changed info_bar on "Reload"
This was caused by an addition I made in the earlier patch to destroy
the info_bar when the user manually reverts or loads a new file).

The callback on revert button was trying to delete it (again) after
the revert _action_ we call triggered a destroy of the widget. Avoid
this by destroying the info_bar before we call the revert action.