gtk: use gtk_action_block_activate() in layer selector
The GHidLayerSelector widget now uses gtk_action_block_activate()
instead of storing/blocking signal IDs. This is the proper way
to suppress signal emission since 2.16.

This also fixes the lockup bug that was supposed to be fixed
by the recent commit 8857757, but still occured when using the
menus to toggle layers.