Fix m4-to-newlib converter filenames
The m4 library uses macros and parameters to describe footprints, which
allows for a *very* flexible footprint naming scheme.  While it is
impossible for pcb to know all valid footprint values, it does have a
list of "standard" ones, and knows how to obtain those.  However, the m4
to newlib converter wasn't using the same footprint naming convention as
everyone else, so produced *.fp files that weren't always useful.  For
example, there were a large number of 300.fp files that all collided.

With this fix, the footprint name matches the value in [...] in the
library window, which is what we tell people to use for the "footprint="
attribute anyway.