Skip the M4 library on Windows
On Windows, set the LibraryContentsCommand to the magic string "*NONE*"
and check for that string while loading the libraries.  This avoids the
windows pcb.exe from trying to run bash or m4 scripts, which are
normally not available, yet allows the user a way to re-enable them if
they want.