action.c: Set PCB = NULL; after freeing it in ActionNew()
When we call CreateNewPCB(), it trips over some code which wants
to access the current PCB in order to determine whether to
auto-save at exit. The code in question does check if PCB is
NULL first, so this is a sufficient fix for this case.

Fixes valgrind output such as:

    ==22404== Invalid read of size 8
    ==22404==    at 0x470944: Parse (parse_l.l:282)
    ==22404==    by 0x471913: ParseFont (parse_l.l:356)
    ==22404==    by 0x447973: CreateDefaultFont (create.c:941)
    ==22404==    by 0x447C58: CreateNewPCB (create.c:211)
    ==22404==    by 0x4273E8: ActionNew (action.c:5902)
    ==22404==  Address 0xd3c4538 is 14,104 bytes inside a block of size 14,120 free'd
    ==22404==    at 0x4C282E0: free (vg_replace_malloc.c:366)
    ==22404==    by 0x4273DE: ActionNew (action.c:5901)

Which is seen when starting an new layout from within PCB.

Affects-bug: lp-856909