Move pcb-printf #if 0 guards for %mr specifier to fix internationalization
Since the previous version of pcb cannot read files with suffixes
on its measures, right now the %mr specifier outputs measures in
cmils, no suffix.

The future behavior would be to select appropriate units and use
those, but for now this is disabled by an #if 0 block around the
relevant call to CoordToString.

This patch moves the #if 0 block to the suffix-specific part of
the code, since we DO want the other behavior of the %mr spec,
specifically, the forcing of '.' as a decimal separator.