toporouter.c: Re-write {coord,point}_intersect_prop() to increase clariry.
Since the wind_* values computed in this function can be -1, 0 or 1,
it is clearer to test them using equality relations, not perform XOR
operations on their bit-patterns.

I believe the old code worked, but it would return an strange constant
for "true". To ram the point home, I've changed the function to return
bool.