Fix parsing of route styles with units
A bug appeared after route styles started being saved with units
suffixed. When loading a PCB file the units were ignored and therefore
it was assumed the values were in cmils.

This was a problem in the get_unit_struct() function which didn't handle
long strings well. If the unit was followed by more characters strcmp
would return non-zero. The function has already worked out the length of
the unit text and therefore strncmp should be used.