Fix locale-dependent gerber output
pcb_printf doesn't have a modifier that lets you select a unit (%mi) and
locale-independent format (like %mr) at the same time.  The gerber
hid was using %mi to output to a file, but that resulted in ',' instead
of '.' for some locales.  This patch switches it to call
g_ascii_formatd() on the value first, to ensure it's output with '.'.