Add ` modifier to pcb-printf, which forces '.' as a decimal separator
This should be used for the gerber exporter and other formats
that need to use '.' regardless of locale in order to obey their
standards.

Note that the %mr specifier (which restricts suffixes to mm and mil
so that old pcb versions can read them) uses this by default, so no
changes need to be made to file.c.