gtk: library window: ignore element if LoadElementToBuffer fails
In gtk/gui-library-window.c, the library window's selection-change
callback calls LoadElementToBuffer() to read the selected element,
so it can show a preview and select the element if necessary.

The return value of this determines whether to set pcb's mode to
PASTEBUFFER_MODE. However, the library window preview is attempted
regardless. If LoadElementToBuffer failed, this causes a NULL
reference and segfault.

This patch skips the preview in the case that LoadElementToBuffer.

Closes-bug: lp-853628
Closes-bug: lp-852600