gtk: allow route style selector to have nothing selected
The affects the the function ghid_route_style_selector_sync() function,
which is called from the RouteStylesChanged action to keep the GUI in
sync with pcb's state. It selects the route style that matches pcb's
state, or if none exists, does nothing.

This patch replaces "does nothing" with "deselects all styles", so that
the GUI will not be misleading.

Closes-bug: lp-699299