Clip polygons after moving between buffers
When moving a polygon between buffers (i.e., from the undo buffer
to the PCB->Data buffer or vice-versa), line- and via-clipping is
not recalculated. This means that any elements drawn on a buffer
before the polygon is moved there, will not clip the polygon.

One way (the only way, AFAIK) to make this happen is by bug
lp-699272, which can reproduced as follows. I have inclided my
explanation of what is happening at each step.

  1. Draw a polygon with a line intersecting it, with clearpoly
     set and visible clearance between the two.a

  2. Delete the polygon. [This moves the polygon verbatim, including
     its clipped form, from PCB->Data to RemoveList, the undo buffer.]

  3. Delete the line. [This moves it from PCB->Data to RemoveList,
     calling RestoreToPolygon() to unclip any polygons on the screen
     (but there aren't any!), then ClearFromPolygon() to clip the
     polygon in the undo buffer (but this is already clipped!). In
     other words, the line is deleted and the polygon unchanged.]

  4. Hit undo. [This moves the line from the undo buffer back to
     PCB->Data, calling RestoreToPolygon() to unclip the polygon in
     the undo buffer, then ClearFromPolygon() to clip any polygons on
     the screen (but there aren't any!). So the polygon is no longer
     clipped.]

  5. Hit undo. [This brings the polygon back, copying it verbatim from
     the undo buffer to PCB->Data. As it is not clipped here, it isn't
     clipped on the screen, giving rise to the "no clipping" bug.]

This patch adds a call to InitClip() to MovePolygonToBuffer(), which
forces a recalculation of a polygon's clipping when it is moved between
buffers. This is perhaps not the most efficient fix, but will prevent
any such synchronization bugs.

Closes-bug: lp-699272