hid/gtk: Fix NetlistChanged action crash when exporting from the commandline.
I broke this in commit 7308f512307158944482227d58e66373fd023d62 when I
added a missing notification to the GUI when free'ing the netlist. It turns
out that a command-line invoked export (such as used to export images in the
doc/ build) will hit this code-path, and provoke the GUI netlist window code
to try and update before the GUI is actually loaded.

For now, lets fix this with a bandaid and just skip the netlist update if the
GUI is not up.

A brief inspection of the code (and test for the doc/ build) suggests that the
Lesstif HID is not affected by this problem. (It has similar code to avoid
creating the netlist window if the main window is not already created).