Action to filter the pastebuffer with an external program.
This patch adds the "Filter(command)" and "! command" actions.
It runs "sh -c command" piping the contents of the pastebuffer to stdin.
If the return code is zero it replaces the buffer with the output of the command.

Both the input and output of the command are saved to temporary files, if pcb is
compiled with debug enabled they are kept, otherwise deleted.

Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>