Add "px" to the units allowed for --rat-thickness
Add "px" to the set of units allowed for option --rat-thickness. The purpose is
to give thickness in terms of a fixed number of screen pixels. Internally, this
special unit is treated like nm. The drawing routine decides to go for fixed
width if rat thickness is below 100nm.

Commit by Andrew Poelstra.
Closes-bug: lp-927422