Add grid-increment-mm (and similar) settings
Added the following settings:

  grid-increment-mm       grid-increment-mil
  size-increment-mm       size-increment-mil
  line-increment-mm       line-increment-mil
  clear-increment-mm     clear-increment-mil

These behave differently than the old settings of the same names (from
20090929 and earlier). They take a measure, not a real number. The -mm
versions affect the increment settings when the user is using a metric
grid, and the -mil versions affect the settings for imperial grids.

Old ~/.pcb/preferences files likely have lines like

  grid-increment-mil = 5

which have been hanging around for several pcb versions. These should
be safely ignored, since, for example, "5" is now read as "5nm", which
is well below the minimum allowed increment setting.

These settings can be read but (so far) not written out, either to the
pcb or to the preferences file.