Fix infinite loop on PCB::skip-drc attribute
When doing a DRC check, pcb skips checking any layers who have the
PCB::skip-drc attribute set. However, when determining whether the
check is complete (via the ListsEmpty()) function, it doesn't skip
these layers.

The result is an infinite loop, since if there is work to be done
on a skip-drc layer, the checker will keep running until it is done
-- i.e., will keep running forever.

This patch skips skip-drc layers in ListsEmpty().

Closes-bug: lp-829123