Fix segfault when routing to non-straight pads
Since the autorouter does not handle non-manhattan lines, it will
not route to non-straight pads. Therefore, when the autorouter
searches for the pins/pads at the endpoints of a rat, it ignores
any non-straght pads, returning NULL from the search.

The autorouter then tries to dereference this pointer and crashes.

The rest of the code, it seems, works to avoid ever giving such
rats to the autorouter, but it can be done if the user explicitly
selects such a rat and selects "Auto-route selected rats".

This patch catches the NULL pointer and simply ignores the rat.

It is possible that this is the cause of bug lp-699255, but this
seems unlikely, since the backtrace on that report does not show
the same NULL dereference, and the reporter was using "10-pin
headers", which presumably have no pads that could be marked as
non-straight.

Affects-bug: lp-699255