Remove manhattan check from rats.c
Since commit 644f8d9, I cannot get the autorouter to segfault when
it attempts to deal with ratlines connected to non-manhattan lines.
So, I see no reason not to draw such lines in the first place.

However, when the autorouter is used, it will ignore any rats which
lead to a non-manhattan line. (And therefore not connect two points,
if the shortest connection would involve connecting diagonal traces.)

This is unexpected behavior, though IMO no more unexpected than the
old behavior, which was to ignore diagonal line segments and simply
route amongst the manhattan ones.

Closes-bug: lp-699248