Fix bom.c behavior regarding xy-in-mm option
If the --xy-in-mm option is passed, that should override any other
unit selection behavior (since this is a legacy option, so anyone
using it is not expecting the new many-unit features).

Also, fix a test-breaking change in output precision.

This has apparently been broken since da403312 (audit bom.c for
the nanometers switchover), but was only noticed recently when the
HID tests were fixed.