gcode HID: implement predrilling
Predrilling is using the milling tool to mark all drill spots. This
eases and enhances accuracy of manual drilling.

This resolves patch #12 and obsoletes patch #11 and #13 of
Traumflug's wishlist:
https://bugs.launchpad.net/pcb/+bug/699497

The difference between patch #12 and this patch is, drills
for predrilling are sorted regardless of size, so performance
should be better.