gcode HID: make G-code output aware of dynamic extents.
This also offsets the produced G-code so the rectangular edge
closest to {0, 0} in pcb is {0, 0} in the G-code. Very useful
when actually producing boards.

The most simple way to achieve this is to draw just the part
inside the outline into the intermediate image. No change to
the PNG tracer needed.

Unfortunately, this exposes a bug, making the traced PNG
image inverted. This will be fixed with the next commit.