Added "pcbignore" to ignore newlib dirs/files.
Added a "pcbignore" file concept, which may be used to
specify the subdirectories and files to be ignored
when constructing the menus of newlib parts in the
"PCB Library" window. Falls back to previous behavior
if no pcbignore files are found. Full documentation is
in a new appendix in doc/pcb.texi (and derivatives).

Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>