hid/gcode: Add a more direct "#include <locale.h>" for setlocale()
The lack of this include caused a failure to build on my machine when
compiled without optimisation (-O0). With any optimisation turned on,
the build would work without the include.

Having chased this strangeness down, it turns out that we were including
libintl.h via global.h, and that libintl.h includes locale.h, IFF
optimisation is turned on at the time it is pre-processed.