find.c: Remove some "flag" variable usage in DRCFind()
Avoid confusion by passing *FLAG constants directly to some functions.
This removes several cases where we would set the variable, use it once,
then change it for the next usage, making the code more succinct.

I was not able to remove the flag variable entirely, as the while loop
calling DoIt() relies on it being changed inside the loop in order to
terminate. It should be possible to re-write the loop exit condition if
desired, but I'm not sure I completely understand the behaviour of the
existing code yet.