src/res_parse.y: handle resource globals gracefully.
Likely, res_file is used somewhere to check something. In case a
resource file was loaded earlier, but the current parsing is from
a string, res_file is still != NULL, so this check would mislead.

Solution: Clear res_file and res_filename as soon as the file
          behind it gets closed.

Also replace a number of zeros with the more appropriate NULL.

Affects-bug: lp-1206882