png: add explicit round() calls to the scaling
After this commit the tests pass on both 32- and 64-bit builds.
Hopefully this time it's really dead; the previous "fix" af27c2d
by myself merely increased the DPI of the output, but this one
attempts to get deterministic rounding behaviour.

Tests from folks with actual 32-bit machines are welcome.

Closes-bug: lp-860037