hid/gcode: Don't use PCB->Extent{Min,Max}{X,Y}.. they are about to be killed
The extent updating code is broken, and adds yet more special-case code which
handles the "outline" layer. Use PCB->MaxWidth and PCB->MaxHeight, these
represent the size the of board, and are the values which the extents code
defaulted to if it did not find any lines on the "outline" layer.