Don't recurse into relatively-specified library directories
When PCB is run from the user's home directory and "." is specified
as a library directory (which may be a valid use case), PCB tries to
read the user's whole home directory.  There is no easy way to tell
if a given relative path should be searched recursively or not, so
don't do it at all.