src/action.c: make Rolands patch for MinMaskGap() complete.
Moved to a more general solution. Instead of looking at the Hole
flag, the mask offset is now always related to the bigger of
Thickness and DrillingHole.

Also applied the same for vias, which are the GUI way of defining
holes.

Last, all this in only applied if a mask hole existed before at
all. Typical non-mounting-hole vias are fully covered by solder
resist.