Write grid size with units.
Apparently this was forgotten earlier.

Reading/parsing code unchanged, so no compatibility issues
expected. Also not considered to be a file format change,
because values are ordered the same, are of the same number
and versions reading this file format must be capable of
reading units already.

This was found when hand-editing a testcase for LP1022076, the
grid size was saved to "funny" numbers, breaking regression tests.