Fix action SetThermal(), implement GetUnitlessValue().
The situation before was double-broken. In opposition to the
documentation in the commandline window, SetThermal() required
a second parameter.

The other broken thing was parsing that value, which is just an
index into a list, but was interpreted like a coordinate given in
centimil.

Both are fixed and on top of this, 'style' is now omittable.
There's a default style in every layout, so no need to stress the
user with finding the number of the wanted style.