hid.h: de-deprecate most SL_* layer types.
They make sense, because not every "layer" drawn to the screen or
written to a file has also a direct pysical representation.

An example would be SL_PDRILL and SL_UDRILL, which are not drawn
by the user, but derived from pins and vias. Other layer types,
like for example the mask layer, exist as both, derived and
physical layer. The physical, user-drawn layer complements the
derived one in such cases.