Remove these almost 25 Makefiles for just copying tests.
Before there was a Makefile/.am/.in in each test directory,
just to define the EXTRA_DIST variable for this directory. A
rather complex approach.

The new strategy is to list all these files in one place, in
tests/Makefile.am. This almost doubles the speed of compilation
after small changes and reduces clutter sent to the command
line a lot, so a great improvement for developers.

Creation of new tests is also simplified, creating a Makefile.am
and editing configure.ac is no longer neccessary.

Previous versions of this commit used wildcards to include new
tests automatically. This strategy was discarded in favour of a
hand-edited list, because some developers aren't exactly excited
by using wildcards.

As we're on the topic, here are a few informative readings
regarding recursive makes and regarding wildcards:

  "Recursive Make Considered Harmful"
  Peter Miller, 2008
  http://aegis.sourceforge.net/auug97.pdf

  GNU automake about recursion:
  http://www.gnu.org/software/automake/manual/html_node/Directories.html

  GNU make about wildcards:
  http://www.gnu.org/software/make/manual/html_node/Wildcard-Function.html

  GNU automake about wildcards:
  http://www.gnu.org/software/automake/manual/html_node/Wildcards.html