Fix the recent addition of smartdisperse.c.
Before committing something, please run a 'make distcheck'. This
is considered to be the most safe way we have to make sure there
is no regression.

If this isn't possible, please build with and without NLS and run
a 'make check' each time. Like:

  ./autogen.sh && ./configure --disable-nls --disable-doc && make
  make check
  ./autogen.sh && ./configure && make
  make check

This is a bit more manual work but gives similar confidence.

All these tests pass now again.