hid/gtk/gui.h: fix definition of LAYER_BUTTON_SILK
LAYER_BUTTON_SILK should give the number of the layer after all copper
layers. Since the copper layers are enumerated starting with zero, this
is equal to the number of copper layers MAX_LAYER.

The previous definition erreously added LAYER_BUTTON_SILK to this number.
This only worked, because src/const.h sets this constant to zero.