Always produce a DRC warn about undected stuff
DRC will catch a near-short, but not a short.  It will catch a
near-break in a trace, but not an actual break.  This is pretty
confusing and potentially harmful to users not familiar with the
behavior, so always produce a DRC entry warning user about the situation
and what to do about it.