Add an elements window
I find that I am often trying to find a component, or a set of
components, on a PCB by refdes.  This patch adds a new window to PCB
that shows a searchable list of all elements.  Elements can be selected
in that list and then optionally selected/deselected.

I am hoping to extend this window to provide access to, and search, the
attributes of elements too.