Moved font libraries into global structures and updated font functions
Put the SystemFontLibrary into the Settings structure and the
EmbeddedFontLibary into the PCB structure. Having the embedded library
as part of the allows the current embedded library data to not get wiped
out if a file fails during parsing.

Also changed some font handling functions to take a font library as a
parameter.