Implemented reading symbols from a font directive
A "Font" directive in a file can now appear by itself, or with
symbols inside it, such as: Font("name of default font") or
Font("name of font") ( symbol data ). Multiple directives are
allowed, so multiple fonts can be stored in one file, or a different
default font can be specified that isn't an included font.