Create new texts with a null font pointer.
The idea being that if you want to completely ignore that the
font system exists, you can. If no objects are created with font
pointers, none will be saved with them. Everything will use
Settings.Font, which is basically the same behavior as previously,
although the pointer is in a different place.