Added --with-included-gettext when building pcb
The w32 build still uses an old gettext (one which builds more readily),
and if we don't apply this option, the intl/ directory in the pcb
source gets replaced with a version which fails to compile.

This lets us use the older gettext tools build, but keep the included
(more up to date) intl library shipped with the pcb tarball.