Fix drawing API wrappers to be static inline
Just inline on its own is wrong, since if the compiler decides NOT to inline
the function, it needs to find an implementation at link-time. Since none of
the places which include this header provide "extern" linkage for the inline
functions. This results in a link error.

Reported-by: Charles Parker <parker.charles@gmail.com>

NB: When this is confirmed to work, I will roll it into the original places
these wrappers were added.