Fixup incorrect usage of MAX_ALL_LAYERS and SILK_LAYER in GUI
These were incorrectly changed to be relative to MAX_ALL_LAYERS, and should be
relateive to MAX_LAYERS like they used to be. This is because we are counting
extra buttons past the number of copper layers (MAX_LAYERS).

Similarly, SILK_LAYER happens to be 2.. this is NOT why there are two buttons
in the GUI after the copper buttons, one is for the current silk layer, one is
for the rats layer.