Make step export cli args work
The '--copper', '--soldermask', '--silk', and '--models' command line
arguments had no effect on the step exporter.  This was because they had
default values of true, and their presence sets their value to true.
This patch changes their defaults to false, thus making toggling
exporting the various parts of the boards possible with the step
exporter.