Use mkstemp() instead of fixed pcb step file name
When exporting a step file, pcb was producing a '_pcb.step' file that
contained the model of the board.  This meant that if there were two
instances of pcb exporting to step file simultaneously, then they would
write over each other's '_pcb.step' file, resulting in broken step file
output.  This patch causes the temporary file name to be generated using
mkstemp().

Note that it uses mkstemp() in an insecure way, as rather than passing
the resulting fd around, it has to pass the filename around.  This can
be fixed later.