pcb.texi: better front matter and copyrights.
By international treaty, the copyright notice for
a book must be either on the title page or on the
back of the title page. When the copyright notice
is on the back of the title page, that page is
customarily not numbered. Therefore, in Texinfo,
the information on the copyright page should be
within @titlepage and @end titlepage commands.

Signed-off-by: bert <bert.timmerman@xs4all.nl>