Rework the tests for ChangeClearSize action.
The tests now use one export to check that clearances are properly
applied to the appropriate layers, and then checks that application of
clearances using an action script.

Note that there is presently a bug in the bbvia code that causes
soldermask clearance to be present on layers that the via does not
intersect. This is the subject of a different bug report.