DRC: Add bloat as a parameter to the IsXInPolygon tests
We might want to do this to some of the other functions too. Also, in
IsLineInPolygon, a calculation was corrected to use 2*bloat instead of
just bloat. Supposedly this breaks something, but I'm not sure what. Now
that bloat is a parameter to these functions, if we figure it out, we
can change it in the function call.