[DRC] Tests for the DRC
Finally! The golden results represent the current state of the pcb DRC,
however, this is not completely correct. There are a number of things
that the DRC doesn't catch that it should. I wanted to get these tests
in place before I started monkeying with the DRC so I'd know if I broke
or fixed anything.

The clearance-misc test is failing the arcs on some platforms, for reasons
undetermined (yet). This file is mostly to remember some ideas for tests
anyway, so, it's not critical.

One note: pcb never resets the object ID counter. So, since all of the
tests are run in the same instance of pcb, the object IDs continue to
increase in each successive test.