[find.c] Eliminate the User variable
This was a global state variable used to indicate that an operation
should be marked as undoable. Now, operations are undoable by default.
To prevent an operation from being registered with the undo system, the
undo system is locked.

Note, there are some potential conflict issues with the way the undo
locks work. There are no checks in place to ensure that a function
doesn't unlock the system if the calling function locked it. For now,
the API isn't used anywhere else, but this should get sorted out before
we start to rely on it too much.