[find.c] Reorganize file and add comments (no code changes)
The file is reorganized to group related definitions and functions. All of
the functions related to the object lists are grouped, all the geometry
functions are grouped, then the connection lookup functions, and finally
the entry points.

The hope is that this will help to minimize the amount of time spent
scrolling around when working on the code.