[DRC] Restructure to use object_list + misc updates
Restructured the DRC to keep lists of violating objects in an
object_list instead of the arrays of object IDs and types. A couple of
other updates happened in the process:
  * set_flag_on_violating_objects added to drc_violation
  * GhidDrcViolation reworked to use DrcViolation instead of duplicating
    the structure.
  * Added DRCReview action. This will mimic the lesstif HID behavior and
    iterate through all of the DRC violations using a pop-up dialog. A
    DRC must be executed first to populate the violation list before
    this can be done. The point here is that next I'm going to make the
    DRC less interactive, but I need to preserve the lesstif behavior.
    This will remove many of the complications with flags and undo
    behavior, and also allow the DRC to be executed using the batch HID
    (without prompting the user for every error!).

There are a few housekeeping items still to attend to. There are some
stray functions that can be removed, probably some variables that no
longer used, headers that can go... that kind of stuff.

The next task on the list is to remove the user interaction from the DRC
and use the DRCReview action for that instead.