[DRC] Check for plow if object doesn't have CLEARLINEFLAG set
When testing if a line/arc/text plows a polygon, there was an early test
of the CLEARLINEFLAG to avoid expensive r_searches if the line is
supposed to connect to the polygon. The problem is that the DRC still
needs to check objects that don't have this flag set as such objects can
still create violations when coupled with the clearance of other
objects.

This will be a performance hit, but I expect it to be negligible on
modern desktops.

This solution was identified by Luis de Arquer.