[DRC] Remove IsPadInPolygon test
This test was added deliberately in commit bc48a574 to fix the DRC of
slanted pads. However, it breaks the DRC for pads that have non-zero
insufficient clearance. A quick test suggests that the test is no longer
necessary.

Note: We'll need to add some tests for slanted pads to the polygon
clearance tests.