[find.c] Make shrunk arc behavior consistent with other objects.
When an object's thickness is < 2 * PCB->Shrink, all tests now treat the
shrunk object as if it has a thickness of zero. The arc tests previously
rejected this case outright.

Note that this behavior misses minimum copper overlap violations
if thickness < 2 * PCB->Shrink and the center line of the object
overlaps with the second object.

Signed-off-by: bert <bert.timmerman@xs4all.nl>