[tests] Split the DRC tests into pieces.
One benefit of this is that the object IDs listed in the DRC report
files will now correspond to the object IDs the objects get when opening
the file directly. Another benefit is that it will be much less
overwhelming of a task to review the smaller files for accuracy.