[DRC] Detect multiple copper too close violations for lines and arcs.
This fixes a bug that prevented the connection lookup code from
detecting multiple copper too close violations. After DRC find flags the
first one, it highlights the rest of the encroaching net so that it's
not reported again. However, this resets the connection lookup lists,
which means that the lookup has to start over from scratch. The flags
were not being properly reset to allow for this, and so on the next
pass, it would see the very first item, see the flags, not add it to the
new list, and then think it's done because there's nothing left in the
list.

The affected tests are primarily the lines and arcs to lines and arcs,
however, the order in which some of the violations are found in the
other tests also changes, so, more files are affected than might be
expected.